pass org name and slug to accept invite
